requirements.txt
aioredis==1.3.0
amqp==2.5.1
asgiref==3.2.3
astroid==2.3.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
billiard==3.6.1.0
boto3==1.10.12
botocore==1.13.12
celery==4.3.0
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
Django==2.2.7
docutils==0.15.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
isort==4.3.21
jmespath==0.9.4
kombu==4.6.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.4.3
python-datauri==0.2.8
python-dateutil==2.8.0
pytz==2019.3
redis==3.3.11
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.6
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.6.0
